refactor toolspane
