+++ src/localpen/app.ts
@@ -37,9 +37,7 @@ import { exportPen } from './export';
 import { createEventsManager } from './events';
 import { getStarterTemplates } from './templates';
 import { defaultConfig } from './config';
-import { createToolsPane } from './tools';
-import { createConsole } from './console';
-import { createCompiledCodeViewer } from './compiled-code-viewer';
+import { createToolsPane, createConsole, createCompiledCodeViewer } from './toolspane';
 import { importCode } from './import';
 import { compress, debounce } from './utils';
 import { getCompiler } from './compiler';
