+++ src/localpen/toolspane/index.ts
@@ -1,0 +1,3 @@
+export * from './compiled-code-viewer';
+export * from './console';
+export * from './tools';
